Refactor WiremockEmrTestHelper and WiremockEmrServerlessTestHelper
