+++ java/clients/src/test/java/sleeper/clients/teardown/ShutdownSystemProcessesIT.java
@@ -49,14 +49,14 @@ import static sleeper.clients.testutil.WiremockCloudWatchTestHelper.disableRuleR
 import static sleeper.clients.testutil.WiremockCloudWatchTestHelper.disableRuleRequestedFor;
 import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.aResponseWithNumRunningApplications;
 import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.aResponseWithNumRunningJobsOnApplication;
+import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.deleteJobsForApplicationsRequest;
 import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.listActiveApplicationRequested;
+import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.listActiveApplicationsRequest;
+import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.listJobsForApplicationsRequest;
+import static sleeper.clients.testutil.WiremockEmrServerlessTestHelper.stopJobForApplicationsRequest;
 import static sleeper.clients.testutil.WiremockEmrTestHelper.aResponseWithNumRunningClusters;
-import static sleeper.clients.testutil.WiremockEmrTestHelper.deleteJobsForApplicationsRequest;
-import static sleeper.clients.testutil.WiremockEmrTestHelper.listActiveApplicationsRequest;
 import static sleeper.clients.testutil.WiremockEmrTestHelper.listActiveClustersRequest;
 import static sleeper.clients.testutil.WiremockEmrTestHelper.listActiveClustersRequested;
-import static sleeper.clients.testutil.WiremockEmrTestHelper.listJobsForApplicationsRequest;
-import static sleeper.clients.testutil.WiremockEmrTestHelper.stopJobForApplicationsRequest;
 import static sleeper.clients.testutil.WiremockEmrTestHelper.terminateJobFlowsRequest;
 import static sleeper.clients.testutil.WiremockEmrTestHelper.terminateJobFlowsRequestWithJobIdCount;
 import static sleeper.clients.testutil.WiremockEmrTestHelper.terminateJobFlowsRequestedFor;
